techstorm-ai/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
openai
pandas
numpy
scikit-learn
google-generativeai
